+++ src/disclosure-manager/_DisclosureContainer.jsx
@@ -10,6 +10,7 @@ import { ApplicationLoadingOverlayProvider } from '../application-loading-overla
 import { NavigationPromptCheckpoint, navigationPromptResolutionOptionsShape, getUnsavedChangesPromptOptions } from '../navigation-prompt';
 import ApplicationErrorBoundary from '../application-error-boundary';
 import BannerCheckpoint from '../banner/_BannerCheckpoint';
+import { addCallback, removeCallback } from './_disclosureCallbacks';
 
 const propTypes = {
   /**
@@ -34,6 +35,15 @@ const DisclosureContainer = injectIntl(({ intl, children, navigationPromptResolu
   const promptCheckpointRef = useRef();
 
   const defaultPromptOptions = useMemo(() => getUnsavedChangesPromptOptions(intl), [intl]);
+
+  useEffect(() => {
+    const callback = disclosureManager.goBack || disclosureManager.closeDisclosure;
+    addCallback(callback);
+
+    return () => {
+      removeCallback(callback);
+    };
+  }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {
     disclosureManager.registerDismissCheck(() => new Promise((resolve, reject) => {