+++ src/terra-dev-site/test/slide-panel-manager/DisclosureComponent.jsx
@@ -5,7 +5,13 @@ import classNames from 'classnames/bind';
 import CollapsibleMenuView from 'terra-collapsible-menu-view';
 import ContentContainer from 'terra-content-container';
 import {
-  availableDisclosureHeights, availableDisclosureWidths, withDisclosureManager, disclosureManagerShape, DisclosureManagerHeaderAdapter,
+  availableDisclosureHeights,
+  availableDisclosureWidths,
+  withDisclosureManager,
+  disclosureManagerShape,
+  DisclosureManagerHeaderAdapter,
+  closeMostRecentDisclosure,
+  getActiveDisclosureCount,
 } from '../../../disclosure-manager';
 import NavigationPrompt from '../../../navigation-prompt';
 import ApplicationLoadingOverlay from '../../../application-loading-overlay';
@@ -39,7 +45,15 @@ class DisclosureComponent extends React.Component {
       hasPendingAction: false,
       hasLoadingOverlay: false,
       hasError: false,
+      disclosureCount: null,
     };
+  }
+
+  componentDidMount() {
+    // Get active disclosure count will not reflect this content until the parent (DiclosureContainer) is mounted.
+    setTimeout(() => {
+      this.setState({ disclosureCount: getActiveDisclosureCount() });
+    });
   }
 
   getId(name) {
@@ -129,7 +143,9 @@ class DisclosureComponent extends React.Component {
 
   render() {
     const { disclosureManager, identifier, renderHeaderAdapter } = this.props;
-    const { hasPendingAction, hasLoadingOverlay, hasError } = this.state;
+    const {
+      hasPendingAction, hasLoadingOverlay, hasError, disclosureCount,
+    } = this.state;
 
     if (hasError) {
       throw new Error('Test Error');
@@ -178,6 +194,7 @@ class DisclosureComponent extends React.Component {
         {disclosureManager && disclosureManager.goBack ? <button type="button" className="go-back" onClick={this.goBack}>Go Back</button> : null}
         {disclosureManager && disclosureManager.maximize ? <button type="button" className="maximize" onClick={this.maximize}>Maximize</button> : null}
         {disclosureManager && disclosureManager.minimize ? <button type="button" className="minimize" onClick={this.minimize}>Minimize</button> : null}
+        <button type="button" className="global-close-disclosure" onClick={closeMostRecentDisclosure}>{`Global Close (${disclosureCount})`}</button>
         {hasPendingAction && <NavigationPrompt description="Test Action" />}
         {hasLoadingOverlay && <ApplicationLoadingOverlay isOpen backgroundStyle="light" />}
       </ContentContainer>